requirements.txt
google-cloud-bigquery>=3.27.0
google-auth>=2.37.0
tqdm>=4.67.1
torchvision==0.17.2
torch==2.2.2
transformers==4.42.3
datasets==2.20.0
open-clip-torch>=2.29.0
huggingface-hub>=0.20.3
pinecone-client==5.0.1
pydantic==2.10.6